# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wa-rev",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/problems/{problem_id}/ai-battles": {
      "post": {
        "tags": [
          "ai_battles"
        ],
        "summary": "Create Ai Battle",
        "operationId": "ai_battles-create_ai_battle",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Problem Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIBattleCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIBattlePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai-battles/{battle_id}": {
      "get": {
        "tags": [
          "ai_battles"
        ],
        "summary": "Read Ai Battle",
        "operationId": "ai_battles-read_ai_battle",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "battle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Battle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIBattlePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai-battles/{battle_id}/generated-code": {
      "post": {
        "tags": [
          "ai_battles"
        ],
        "summary": "Save Generated Code",
        "operationId": "ai_battles-save_generated_code",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "battle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Battle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIBattleGeneratedCodeCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIBattleParticipantPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIBattleCreate": {
        "properties": {
          "contest_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contest Id"
          },
          "user_submission_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Submission Id"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "user_submission_id",
          "models"
        ],
        "title": "AIBattleCreate"
      },
      "AIBattleGeneratedCodeCreate": {
        "properties": {
          "participant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Participant Id"
          },
          "model_id": {
            "type": "string",
            "maxLength": 255,
            "title": "Model Id"
          },
          "source_code": {
            "type": "string",
            "minLength": 1,
            "title": "Source Code"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason"
          },
          "usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage"
          }
        },
        "type": "object",
        "required": [
          "participant_id",
          "model_id",
          "source_code"
        ],
        "title": "AIBattleGeneratedCodeCreate"
      },
      "AIBattleParticipantPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "battle_id": {
            "type": "string",
            "format": "uuid",
            "title": "Battle Id"
          },
          "participant_type": {
            "type": "string",
            "title": "Participant Type"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          },
          "generation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Status"
          },
          "generation_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Error"
          },
          "order_num": {
            "type": "integer",
            "title": "Order Num"
          }
        },
        "type": "object",
        "required": [
          "id",
          "battle_id",
          "participant_type",
          "display_name",
          "order_num"
        ],
        "title": "AIBattleParticipantPublic"
      },
      "AIBattlePublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "problem_id": {
            "type": "string",
            "format": "uuid",
            "title": "Problem Id"
          },
          "contest_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contest Id"
          },
          "user_submission_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Submission Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/AIBattleParticipantPublic"
            },
            "type": "array",
            "title": "Participants",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "problem_id",
          "user_submission_id",
          "status"
        ],
        "title": "AIBattlePublic"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}